minor refactoring/polishing
