+++ src/pages/UsersPage/pages/SingleUserPage/sections/Posts/features/Post/components/index.ts
