+++ src/pages/UsersPage/pages/SingleUserPage/sections/index.ts
@@ -1,3 +1,3 @@
 export {default as UserBreadcrumb} from './UserBreadcrumb';
-export {default as UserDetails} from './UserDetails/UserDetails';
+export {default as UserDetails} from './UserDetails';
 export {default as Posts} from './Posts/Posts'
